refactor: separate Comment.tsx
